+++ CHANGELOG.md
@@ -23,3 +23,4 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Bug Fixes
  - Fixed a possible crash on shutdown when using levelDb.
+ - Set an idle timeout for metrics connections, to clean up ports when no longer used